pyproject.toml
[project]
name = "wb-managerbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles==24.1.0",
    "aiogram==3.17.0",
    "aiogram-i18n==1.4",
    "aiohappyeyeballs==2.4.4",
    "aiohttp==3.11.11",
    "aiomysql==0.2.0",
    "aiosignal==1.3.2",
    "aiosqlite==0.21.0",
    "alembic==1.14.1",
    "annotated-types==0.7.0",
    "asyncpg==0.30.0",
    "attrs==24.3.0",
    "babel==2.16.0",
    "cachetools==6.1.0",
    "certifi==2024.12.14",
    "cffi==1.17.1",
    "click==8.1.8",
    "cryptography==45.0.5",
    "fluent-runtime==0.4.0",
    "fluent-syntax==0.19.0",
    "frozenlist==1.5.0",
    "greenlet==3.1.1",
    "hiredis==3.1.0",
    "idna==3.10",
    "levenshtein==0.27.1",
    "magic-filter==1.0.12",
    "mako==1.3.8",
    "markupsafe==3.0.2",
    "msgpack==1.1.1",
    "msgspec==0.19.0",
    "multidict==6.1.0",
    "numpy==2.3.2",
    "opencv-python==4.11.0.86",
    "packaging==25.0",
    "pillow==11.3.0",
    "propcache==0.2.1",
    "psutil==7.0.0",
    "pyaes==1.6.1",
    "pyasn1==0.6.1",
    "pycparser==2.22",
    "pydantic==2.10.5",
    "pydantic-core==2.27.2",
    "pydantic-settings==2.7.1",
    "pymysql==1.1.1",
    "pytesseract==0.3.13",
    "python-dotenv==1.0.1",
    "pytz==2024.2",
    "rapidfuzz==3.13.0",
    "redis==5.2.1",
    "rsa==4.9.1",
    "sqlalchemy==2.0.37",
    "telethon==1.40.0",
    "typing-extensions==4.12.2",
    "typing-inspection==0.4.1",
    "tzdata==2024.2",
    "uvloop==0.21.0",
    "yarl==1.18.3",
]
